Replaces BigDecimal representation of overtime duration with Duration type
BigDecimal representation (faction is used as minutes) leads to implicit
mapping instead of clear representation.

This commit does not affect, commit only adds conversion
* Application hours
* Overtime settings (min, max)

see #1673

+++ src/main/java/org/synyx/urlaubsverwaltung/overtime/OvertimeRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.synyx.urlaubsverwaltung.person.Person;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -19,7 +18,7 @@ interface OvertimeRepository extends CrudRepository<Overtime, Integer> {
     List<Overtime> findByPerson(Person person);
 
     @Query("SELECT SUM(overtime.duration) FROM Overtime overtime WHERE overtime.person = :person")
-    BigDecimal calculateTotalHoursForPerson(@Param("person") Person person);
+    Double calculateTotalHoursForPerson(@Param("person") Person person);
 
     @Query(
         "SELECT overtime FROM Overtime overtime WHERE overtime.person = :person "
